+++ build_runner/CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.8.4
 
 - Log the number of completed actions on successful builds.
-- Deprecated `PostProcessBuilder` and `PosProcessBuildStep`. These should be
-  import from `package:build` instead.
+- Support the new `isRoot` field for `BuilderOptions` so that builders can
+  do different things for the root package.
+- Deprecated `PostProcessBuilder` and `PostProcessBuildStep`. These should be
+  imported from `package:build` instead.
 
 ## 0.8.3